LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-img-element/image-loading-lazy-in-script-disabled-iframe-expected.txt

CONSOLE MESSAGE: Blocked script execution in 'http://localhost:8800/html/semantics/embedded-content/the-img-element/resources/image-loading-lazy-in-viewport.html' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.


PASS Images with loading='lazy' in script disabled iframe are not handled
       as 'lazy' 
